requirements.txt
mrjob
warc
simplejson
boto3
sphinx_bootstrap_theme
sqlalchemy
psycopg2